Move game utils for mc to the strategies folder
